-Werror signed/unsigned

+++ lib/cpp-test/solver/toy_dataset.ipp
@@ -18,7 +18,7 @@ SArrayDouble2dPtr get_features() {
                             0.12,  0.00,  -0.00, -1.33, -0.00, 0.85,  3.03};
 
   ArrayDouble2d features(n_samples, n_features);
-  for (int i = 0; i < features_data.size(); ++i) {
+  for (size_t i = 0; i < features_data.size(); ++i) {
     features[i] = features_data[i];
   }
   return features.as_sarray2d_ptr();
